requirements.txt
openai
python-dotenv
fastapi
fastapi-cors
uvicorn
azure-search-documents==11.4.0b3
azure-storage-blob